Alter one of the possible AI names Twosre -> Twosret
Brought to our attention by @liulitchi

Fixes #26

----

The list of AI Names for the Egyptian civ are based on a list provided in the "New Kingdom Egypt Civilization Profile.pdf" design document written by @zophim. (https://wildfiregames.com/forum/index.php?/topic/18458-collection-design-docs/&do=findComment&comment=291582)

On page 23 of this document, we have:
> Twosre [Thuoris - female Pharoah]

Several online sources provide alternate variations on the name of what appear to be the same individual, including:
* "Twosret"
* "Twosre"
* "Thuoris"
* "Tausert"
* "Tawosret"
* "Tawesret"

Of which "Twosret" appears to be the most commonly used variant.